Simplified the design to fit YouTube Size constraints
YouTube can be viewed on different devices and each one is supposed to have a
different size (although when actually testing this, it seems that YouTube uses
pretty much the same part of the banner image regardless of media).

Initially I just created a design with the Logo's, however, this felt a little
too simplified and would not have said anything to those who didnt recognise the
logos.

The design was refactored to place the names the logos represented under each
logo.  The font used was Ubuntu as its an elegant font to use.  The colour of
each name was taken from the predominant colour of the logo it describes.  There
was not much difference between the Green and Blue of the Clojure logo but Green
seemed to work better for the text.