Modified design to include ClojureScript too
Added CLJS logo from https://github.com/cljs/logo

Created a new logo to represent Figwheel, as there doesnt seem to be an official
logo.  The idea behind the logo is to use the fact that a fig cut in half is
roughly the shape of a heart (especially when held upside down).  As Figwheel is
for ClojureScript, then its incorporated into the existing logo above.

Added an updated image of the YouTube channel with the new banner applied.