New CIDER logo
Added a border around the logo to covert the logo to a round shape.  Added
arrows on the border to express the an iterative nature of development with
CIDER.

Initially used the Clojure colours for the border, however, the cider colours
are visually more appealing and keep the colour scheme simpler.